Removed unused instructions and imports We will worry about reloading widgets later
The KvFileInstruction is redundant. The user can just open and read the kvfile themselves, and use KvStrInstruction. This will require passing more data between processes, but most kv files are relatively short. If we come across performance issue in the future than we can update to saving the kv file, and passing the filepath.